requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Google API
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Ollama client
ollama>=0.4.0
httpx>=0.27.0

# Async support
nest-asyncio>=1.5.0

# Metrics and monitoring
prometheus-client>=0.19.0
psutil>=5.9.0
